+++ bin/router/src/pipeline/mod.rs
@@ -1,7 +1,11 @@
-use std::{borrow::Cow, sync::Arc};
+use std::sync::Arc;
 
-use hive_router_plan_executor::execution::plan::PlanExecutionOutput;
-use hive_router_query_planner::utils::cancellation::CancellationToken;
+use hive_router_plan_executor::execution::plan::{
+    ClientRequestDetails, OperationDetails, PlanExecutionOutput,
+};
+use hive_router_query_planner::{
+    state::supergraph_state::OperationKind, utils::cancellation::CancellationToken,
+};
 use http::{header::CONTENT_TYPE, HeaderValue, Method};
 use ntex::{
     util::Bytes,
@@ -12,7 +16,7 @@ use crate::{
     pipeline::{
         coerce_variables::coerce_request_variables,
         csrf_prevention::perform_csrf_prevention,
-        error::PipelineError,
+        error::{PipelineError, PipelineErrorFromAcceptHeader, PipelineErrorVariant},
         execution::execute_plan,
         execution_request::get_execution_request,
         header::{
@@ -53,9 +57,13 @@ pub async fn graphql_request_handler(
     schema_state: &Arc<SchemaState>,
 ) -> web::HttpResponse {
     if req.method() == Method::GET && req.accepts_content_type(*TEXT_HTML_CONTENT_TYPE) {
-        return web::HttpResponse::Ok()
-            .header(CONTENT_TYPE, *TEXT_HTML_CONTENT_TYPE)
-            .body(GRAPHIQL_HTML);
+        if shared_state.router_config.graphiql.enabled {
+            return web::HttpResponse::Ok()
+                .header(CONTENT_TYPE, *TEXT_HTML_CONTENT_TYPE)
+                .body(GRAPHIQL_HTML);
+        } else {
+            return web::HttpResponse::NotFound().into();
+        }
     }
 
     if let Some(jwt) = &shared_state.jwt_auth_runtime {
@@ -102,12 +110,11 @@ pub async fn execute_pipeline(
 ) -> Result<PlanExecutionOutput, PipelineError> {
     perform_csrf_prevention(req, &shared_state.router_config.csrf)?;
 
-    let execution_request = get_execution_request(req, body_bytes).await?;
+    let mut execution_request = get_execution_request(req, body_bytes).await?;
     let parser_payload = parse_operation_with_cache(req, shared_state, &execution_request).await?;
     validate_operation_with_cache(req, supergraph, schema_state, shared_state, &parser_payload)
         .await?;
 
-    let progressive_override_ctx = request_override_context()?;
     let normalize_payload = normalize_request_with_cache(
         req,
         supergraph,
@@ -116,12 +123,33 @@ pub async fn execute_pipeline(
         &parser_payload,
     )
     .await?;
-    let query = Cow::Owned(execution_request.query.clone());
     let variable_payload =
-        coerce_request_variables(req, supergraph, execution_request, &normalize_payload)?;
+        coerce_request_variables(req, supergraph, &mut execution_request, &normalize_payload)?;
 
     let query_plan_cancellation_token =
         CancellationToken::with_timeout(shared_state.router_config.query_planner.timeout);
+
+    let client_request_details = ClientRequestDetails {
+        method: req.method(),
+        url: req.uri(),
+        headers: req.headers(),
+        operation: OperationDetails {
+            name: normalize_payload.operation_for_plan.name.as_deref(),
+            kind: match normalize_payload.operation_for_plan.operation_kind {
+                Some(OperationKind::Query) => "query",
+                Some(OperationKind::Mutation) => "mutation",
+                Some(OperationKind::Subscription) => "subscription",
+                None => "query",
+            },
+            query: &execution_request.query,
+        },
+    };
+
+    let progressive_override_ctx = request_override_context(
+        &shared_state.override_labels_evaluator,
+        &client_request_details,
+    )
+    .map_err(|error| req.new_pipeline_error(PipelineErrorVariant::LabelEvaluationError(error)))?;
 
     let query_plan_payload = plan_operation_with_cache(
         req,
@@ -135,12 +163,12 @@ pub async fn execute_pipeline(
 
     let execution_result = execute_plan(
         req,
-        query,
         supergraph,
         shared_state,
         &normalize_payload,
         &query_plan_payload,
         &variable_payload,
+        &client_request_details,
     )
     .await?;