+++ lib/query-planner/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.2](https://github.com/graphql-hive/router/compare/hive-router-query-planner-v2.0.1...hive-router-query-planner-v2.0.2) - 2025-10-27
+
+### <!-- 1 -->Bug Fixes
+
+- *(query-planner, router)* fix introspection for federation v1 supergraph ([#526](https://github.com/graphql-hive/router/pull/526))
+
+## [2.0.1](https://github.com/graphql-hive/router/compare/hive-router-query-planner-v2.0.0...hive-router-query-planner-v2.0.1) - 2025-10-23
+
+### Fixed
+
+- *(query-planner)* inline nested fragment spreads ([#502](https://github.com/graphql-hive/router/pull/502))
+
 ## [2.0.0](https://github.com/graphql-hive/router/compare/hive-router-query-planner-v1.0.0...hive-router-query-planner-v2.0.0) - 2025-10-05
 
 ### Other